Check formula exists before processing dependency
Before processing any of the dependencies listed in a Brewfile,
brewdler will now check that the currently installed version of
homebrew includes a formula for every single dependency.

This avoids confusing error messages downstream when brewdler is
trying to execute the `brew list` or the `brew info` command for
a dependency that doesn't have a corresponding brew formula.