Massively speedup `brew bundle check`.
The main slowdown with `brew bundle` check was the repeated shelling out
to other `brew` commands unnecessary. We can't avoid this for
`brew cask` (for now) but we can do so for the others. Additionally
now that `brew bundle` and `brew services` are core commands we don't
need to check if they are installed and we don't need to check if
Homebrew is installed (as it's require for `brew bundle` to work).

As well all this: rename specs to be more consistent.