Require bundler explicitly
Currently, when we want to run the test suite, we must run `bundle exec
rake spec`. When run `rake spec` alone, `without bundler exec`, an
exception is thrown `NameError: uninitialized constant Bundler` because
we when run with `bundle exec`, `bundler` is automated required for us.
This a confused behaviour because in the code, we don't use bundler.
And in fact, if we remove all the bundler stuffs, then just run `rake
spec`, the test will pass.

So i may some changes in the `spec_helper`. First, require `bundler`
explicitly. Second, run all specs under `Bundler.with_clean_env`, which
make the test easy to understand and separate setup test env with the
actual test.