formula_upgradable? was always returning false
The input to formula_upgradable? is a formula name, so the check
formula_in_array?(formula, upgradable_formulae) was always failing
because upgradable_formulae was an array of formulae, not an array of
formula names. This commit changes outdated_formulae and pinned_formulae
to each return an array of formula names rather than an array of
formulae, with the result that upgradable_formulae, which is defined as
the difference between outdated_formulae and pinned_formulae, will now
always be an array of formula names, and the check
formula_in_array?(formula, upgradable_formulae) will function as
expected.

Closes #144