Ensures rspec 3.5.0 is used
I encountered an error saying that `filter_run_when_matching` was
missing on the RSpec config object when running specs for #324. It turns
out that this method was added in rspec 3.5.0 and I had an older
version of RSpec installed and that older version met the requirement in
the Gemfile of "any version!"

Gemfile.lock is preferrable compared to adding the dependency version in
Gemfile.

https://github.com/Homebrew/homebrew-bundle/pull/325#issuecomment-364416356