Fixes check test by accounting for single parse of Brewfile
Previously, each category tested (brew, cask, tap, mas) could parse the
Brewfile independently through a memoized line in each. This seemed
unnecessary, so I moved that parse to `Bundle::Commands::Check#run`.
This breaks this test because the methods invoked in this test were the
ones that previously executed that parse!