Corrects check dependency satisfaction test, adds empty Brewfile check test
I realized in the process of writing the previous commit message that my
dependency satisfaction test didn't actually test what it said it
tested. Rather, it tested that the dependencies were unspecified. That
is, an empty Brewfile. For the unlikely event of an empty Brewfile,
check probably should return that there is no work to be performed.