Ensure `MacAppStoreDumper` only dumps valid app details
During a routine `brew bundle` I came across the following exception
(which manages to halt all commands):

```
Error: Invalid Brewfile: uninitialized constant #<Class:#<Bundle::Dsl:0x00000101a5d898>>::Install
```

Searching around the issue tracker, I noticed a couple of existing[1]
issues[2] where this was occurring when the `mas list` output included
an application from the app store that wasn't formatted as `brew bundle`
expected. Sure enough, this was also my issue.

```
497799835 Xcode (9.2)
409183694 Keynote (8.1)
408981434 iMovie (10.1.8)
 Install macOS High Sierra (13105)
409201541 Pages (7.1)
682658836 GarageBand (10.3.1)
409203825 Numbers (5.1)
```

Taking a look at the `MacAppStoreDumper` class, the way this output was
previously generated was by splitting on the first bit of whitespace it
encounters and assigned the two parts to the app ID and app name
respectively. In the case of the High Sierra upgrade app, there wasn't
an ID for it and the first bit of whitespace came after "Install". This
explains why the generated line would be:

```
mas "macOS High Sierra", id: Install
```

It was splitting on the first bit of whitespace it encountered and
didn't restrict what that first value could possibly be for it to be a
valid application.

In order to firm up the app details extraction method a little more,
I've swapped the extraction to use a regular expression with named
captures.  By doing this, it introduces a (very) loose validation of
what the `mas list` output should be providing. If the output format
changes, it won't break `brew bundle` subcommands, it will just fail to
add them to the generated output.

Additionally, there is a new `compact` call added to the `@apps` value
to ensure that the `dump` method doesn't need to concern itself with
iterating on empty values and potentially throwing `nilClass` exceptions
for the `downcase` name calls.

Fixes #308
Fixes #321

[1]: https://github.com/Homebrew/homebrew-bundle/issues/321
[2]: https://github.com/Homebrew/homebrew-bundle/issues/308