Extend valid `mas` app name regex
#349 introduced a basic regex for determining if the app names provided
via `mas` were valid and would discard any that failed to meet this
requirement. Apple are quite generous[1] when it comes to what you can
name an application and in essence, the only restriction they place on
it is that it cannot infringe on a trademark and must be under 30
characters long.

This means that there are real applications that fail to meet the
initial regex of `\w\s` due to using non-alphanumeric characters.
"Pastebin It!" is an example of this.

To account for this, I've extended the regex to accept any character
for the `name` portion of the regex. This might look fragile (and I
would agree!) however it is the safest way to ensure that we don't
silently drop applications from being included in the output. Should
Apple tighten this restriction in the future, we can firm it up with a
better regex.

Fixes #359

[1]: https://developer.apple.com/app-store/review/guidelines/#accurate-metadata