add `brew bundle --file=path`
I was looking for what the flag was for this (or whether it was an option at all) and had to read the man pages to figure out. It doesn't show up easily after googling either so I thought it would be helpful to add a note in the readme. Let me know what y'all think it's helpful and whether I should rephrase the note. 

Thanks!