Refactors Bundle::Installer to separate tasks
The install method had grown to an unwieldy complexity. Mike
suggested[1] that it be split to encapsulate exception handling, but
I decided to expand past that into breaking out error reporting, too.
The result is significantly more readable and ultimately testable.

Future work might include introducing an InstallerTaskResult struct to
replace the Hash with 1 or a Hash with the error encountered. I don't
like that convention!

This may affect early exit — exit on first error — so there may be some
improvement to handle that desired feature.

[1]: https://github.com/Homebrew/homebrew-bundle/pull/457/files#r263770108