Add support for Whalebrew to bundle
This introduces support for Whalebrew, a slightly different approach
to managing packages within Homebrew into homebrew-bundle. The
reasoning behind this is outlined in #655 but in short, if you manage
your environment with `brew` packages, it becomes increasingly
difficult to use this tool alongside it as you need to manage the
Brewfile _and_ a custom wrapper for installing Whalebrew packages
(can be done, just needs handrolled Ruby). By supporting this in
`homebrew-bundle` I'm hoping it allows these two installation methods
to co-exist and not step on each other.

Closes #655