exec: Allow the use of commands not in the PATH
This extends `brew bundle exec` with the ability to run a command which isn't in the path, for example, my use case was running a `./configure` script using the Brewfile magic, but this wasn't supported.

My patch basically just skips the storage of the path to the command in `command_path` as well as the related null if the command has a directory separator in it anywhere, and then uses `exec` as normal.

In effect, this brings it up to acting more like a shell would, meaning you can run a configure script via `brew bundle exec` as `./configure`, but not as `configure`. On the other hand, you can now run something like `brew bundle exec bin/setup`, or pass an absolute path to an executable.

Specs have been added for the three types of paths this should interact with; absolute and relative paths, and relative paths which do not contain a `.` indicating the current directory.