Improve non-verbose output
When installing the verb (i.e. "installing", "tapping") is output after
the installation is complete (rather than before). This means if an
installation is taking a long time or frozen completely there's nothing
in the (non-verbose) output to indicate which install is taking the
time.

Instead, let's add a "preinstall" step before "install" for each
installer which identifies whether an installation will be skipped or
not. This will block output but is quicker (as it never does the
actual installation but will do a network call for e.g. `mas`) and
outputs `Using` _after_ the skip check but `Installing` _before_
actually beginning the installation.