Emit brew install args when in verbose mode
I'm setting up a Brewfile that will install the dependencies for a package that is in Homebrew but we're needing to install the package itself outside of Homebrew. I added `args: ['only-dependencies']` to the brew line but it wasn't working: the package itself was getting installed. So, I added this logging facility and realized that I'd typed "only_dependencies" instead. It seems prudent to show all args passed to `brew install` when verbosity is requested.