exec: support `nodenv`, `pyenv` and `rbenv`
If `nodenv`, `pyenv`, or `rbenv` are in the `Brewfile`, let's add their
`shims` directory to `PATH` during `brew bundle exec`.

This makes it so that `brew bundle exec` just works without relying on
the user adding the required `node`/`python3`/`ruby` formula to their
`Brewfile` and then doing `brew *env-sync`.