Add `postinstall` support for `brew` and `cask`
This allows `postinstall` to be used to run a command after a formula
or cask is installed.

While we're here, also adjust the `restart_service` behaviour to
correctly match the comment i.e. require `always` to restart every
time and otherwise just restart on an install or upgrade of a formula.